Reorganise FFL application layer into domain-concern files
Split Commands methods across squad.go, team.go, score.go, reactions.go
by reason-to-change rather than interaction type. commands.go now holds
only the struct and constructor. Renamed import.go → dataops.go.
