+++ services/ffl/internal/application/dataops.go
@@ -188,4 +188,3 @@ func (c *DataOpsCommands) MarkTeamFinal(ctx context.Context, params MarkTeamFina
 	}
 	return nil
 }
-
